+++ client/webpack.config.js
@@ -76,13 +76,6 @@ function prismJsLicenseTransform(packages) {
   return genericTransform(packages, (pkg) => pkg.name.startsWith("prismjs"));
 }
 
-function altchaLicenseTransform(packages) {
-  return genericTransform(
-    packages,
-    (pkg) => pkg.name.startsWith("altcha") || pkg.name.startsWith("@altcha"),
-  );
-}
-
 function bootstrapLicenseTransform(packages) {
   return genericTransform(packages, (pkg) => pkg.name.startsWith("bootstrap"));
 }
@@ -125,6 +118,20 @@ function tomSelectLicenseTransform(packages) {
     return `SPDX-FileCopyrightText: ${author}\n\nSPDX-License-Identifier: ${pkg.license}`;
   }
   return "";
+}
+function altchaLicenseTransform() {
+  const pkg = require(
+    path.join(__dirname, "node_modules", "altcha", "package.json"),
+  );
+  const author =
+    typeof pkg.author === "string"
+      ? pkg.author
+      : pkg.author?.email
+        ? `${pkg.author.name} <${pkg.author.email}>`
+        : pkg.author?.name
+          ? pkg.author.name
+          : "";
+  return `SPDX-FileCopyrightText: ${author}\n\nSPDX-License-Identifier: ${pkg.license}`;
 }
 // REUSE-IgnoreEnd