+++ weblate/trans/forms.py
@@ -2497,7 +2497,6 @@ class ProjectSettingsForm(SettingsBaseForm, ProjectDocsMixin, ProjectAntispamMix
             "name",
             "web",
             "instructions",
-            "set_language_team",
             "use_shared_tm",
             "contribute_shared_tm",
             "autoclean_tm",
@@ -2592,15 +2591,6 @@ class ProjectSettingsForm(SettingsBaseForm, ProjectDocsMixin, ProjectAntispamMix
                 }
             )
 
-    def save(self, commit: bool = True) -> None:
-        super().save(commit=commit)
-        if self.changed_access:
-            self.instance.change_set.create(
-                action=ActionEvents.ACCESS_EDIT,
-                user=self.user,
-                details={"access_control": self.instance.access_control},
-            )
-
     def __init__(self, request: AuthenticatedHttpRequest, *args, **kwargs) -> None:
         super().__init__(request, *args, **kwargs)
         self.user = request.user
@@ -2645,7 +2635,6 @@ class ProjectSettingsForm(SettingsBaseForm, ProjectDocsMixin, ProjectAntispamMix
                 ),
                 Tab(
                     gettext("Workflow"),
-                    "set_language_team",
                     "use_shared_tm",
                     "contribute_shared_tm",
                     "autoclean_tm",